# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smit CSC Info API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /membership/subscribe:
    post:
      operationId: subscribeMembership
      tags: [membership]
      summary: Initiate a membership subscription (payment)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeBody"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SubscribeBody:
      type: object
      properties:
        planId:
          type: string
      required:
        - planId
...
    PaymentInitResponse:
      type: object
      properties:
        transactionId:
          type: string
        redirectUrl:
          type: string
        amount:
          type: integer
        plan:
          type: string
      required:
        - transactionId
        - redirectUrl
        - amount
        - plan
...